+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Add `renderDonation` on `RewardCard` component.
+- Feature: Add `k-Label--micro` class.
 
 ## [20.12.3] - 2018-05-22