+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `ContributionCard` and `DonationCard` components.
 - Feature: Add `renderDonation` on `RewardCard` component.
 - Feature: Add `k-Label--micro` class.