+++ CHANGELOG.md
@@ -4,7 +4,37 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: Add `paragraph` props on `ProjectCard` component.
+- Feature: Add `renderDonation` on `RewardCard` component.
+- Feature: Add `k-Label--micro` class.
+
+## [20.12.3] - 2018-05-22
+
+- Fix: Remove useless props on `RewardCard` component which ones
+  are: `imageSrcSmall`, `imageSrc` and fix image height to keep original ratio.
+
+## [20.12.2] - 2018-05-18
+
+Fix:
+- Fix font & missing prop in `RewardCard` component.
+
+## [20.12.1] - 2018-05-17
+
+Fix:
+- Fix `RewardCard` custom style.
+
+## [20.12.0] - 2018-05-17
+
+Feature:
+- Add `buttonOnMouseEnter`, `buttonOnMouseLeave`, `buttonOnClick` and custom
+  props on `RewardCard` component.
+
+## [20.11.0] - 2018-05-16
+
+Feature:
+- Add `paragraph` props on `ProjectCard` component.
+
+Fix:
+- Add `title` and `myContribution` early return and fix bug with the grid.
 
 ## [20.10.0] - 2018-05-04